Fix build bustage when compiler warnings as errors (int32_t/uint32_t)

+++ image/src/VectorImage.cpp
@@ -841,7 +841,7 @@ VectorImage::CreateSurfaceAndShow(const SVGDrawingParameters& aParams)
   // x or y > maxDimension, because for vector images this can cause bad perf
   // issues if large sizes are scaled repeatedly (a rather common scenario)
   // that can quickly exhaust the cache.
-  uint32_t maxDimension = 400;
+  int32_t maxDimension = 400;
   
   bool bypassCache = bool(aParams.flags & FLAG_BYPASS_SURFACE_CACHE) ||
                      // Refuse to cache animated images:
@@ -850,8 +850,8 @@ VectorImage::CreateSurfaceAndShow(const SVGDrawingParameters& aParams)
                      // The image is too big to fit in the cache:
                      !SurfaceCache::CanHold(aParams.size) ||
                      // Image x or y is larger than our cache cap:
-                     unsigned(aParams.size.width) > maxDimension ||
-                     unsigned(aParams.size.height) > maxDimension;
+                     aParams.size.width > maxDimension ||
+                     aParams.size.height > maxDimension;
   if (bypassCache)
     return Show(svgDrawable, aParams);
